+++ docs.bruno/Testing_2.bru
@@ -1,0 +1,11 @@
+meta {
+  name: Testing_2
+  type: http
+  seq: 4
+}
+
+get {
+  url: http://localhost:3000/
+  body: none
+  auth: none
+}
