+++ docs.bruno/Testing_3.bru
@@ -1,0 +1,11 @@
+meta {
+  name: Testing_3
+  type: http
+  seq: 5
+}
+
+get {
+  url: http://localhost:3000/ping
+  body: none
+  auth: none
+}
