+++ docs.bruno/apidata.bru
@@ -1,0 +1,17 @@
+meta {
+  name: apidata
+  type: http
+  seq: 3
+}
+
+post {
+  url: https://somehow-south.codedamn.pro/api/data
+  body: json
+  auth: none
+}
+
+body:json {
+  {
+    "best_ug_degree": "kalvium"
+  }
+}
