+++ docs.bruno/data.bru
@@ -1,0 +1,11 @@
+meta {
+  name: data
+  type: http
+  seq: 2
+}
+
+get {
+  url: https://tape-highway.codedamn.app/data
+  body: none
+  auth: none
+}
