+++ package.json
@@ -1,7 +1,19 @@
 {
+  "name": "calf-kings",
+  "version": "1.0.0",
+  "description": "List of footballers with impressive calves",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon server.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
-    "mongoose": "^8.2.0"
+    "mongoose": "^8.2.0",
+    "nodemon": "^3.1.0"
   }
 }
